Move App to HomePage
