+++ notes.md
@@ -135,6 +135,7 @@ Vragenlijst
 ## Als een gebruiker wil ik verschillende paginas kunnen bekijken
 
 - [x] Om de hele app moet een Router heen komen
+- [x] Inhoud van App.js verplaatsen naar HomePage
 - [ ] Switch, Routes nodig voor de verschillende paginas
 - [ ] Links nodig om te navigeren
